now DELETE works properly

+++ doctor-manager-working/src/main/java/com/dziobsoft/doctormanager/doctor/dao/VisitDao.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface VisitDao {
     List<Visit> getList(LocalDate fromDate, LocalDate toDate);
     void persist(Visit visit);
-    void deleteVisit(int id);
-    Visit getVisitById(int id);
+    void deleteVisit(long id);
+    Visit getVisitById(long id);
 }
