+++ doctor-manager-working/src/main/java/com/dziobsoft/doctormanager/doctor/services/VisitService.java
@@ -10,5 +10,5 @@ public interface VisitService {
     List<Visit> getList(LocalDate fromDate, LocalDate toDate);
     Boolean isValid(VisitResource visitResource);
     void putVisit(Visit visit);
-    void deleteVisit(int id);
+    void deleteVisit(long id);
 }
